+++ src/main/java/io/github/smokahs/culinarycompat/recipe/MultiCuttingExtras.java
@@ -1,0 +1,7 @@
+package io.github.smokahs.culinarycompat.recipe;
+
+import net.minecraftforge.items.ItemStackHandler;
+
+public interface MultiCuttingExtras {
+	ItemStackHandler culinarycompat$getExtras();
+}