+++ src/main/java/io/github/smokahs/culinarycompat/registry/Recipes.java
@@ -1,17 +1,41 @@
 package io.github.smokahs.culinarycompat.registry;
 
 import java.util.function.Predicate;
+import java.util.function.Supplier;
 
 import net.minecraft.core.NonNullList;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.inventory.CraftingContainer;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.CraftingBookCategory;
 import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.crafting.RecipeSerializer;
+import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.item.crafting.ShapelessRecipe;
+import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.registries.RegistryObject;
+
+import io.github.smokahs.culinarycompat.CulinaryCompat;
+import io.github.smokahs.culinarycompat.recipe.MultiCutting;
 
 public final class Recipes {
+	public static final DeferredRegister<RecipeSerializer<?>> RECIPE_SERIALIZERS = DeferredRegister
+			.create(ForgeRegistries.RECIPE_SERIALIZERS, CulinaryCompat.MODID);
+	public static final DeferredRegister<RecipeType<?>> RECIPE_TYPES = DeferredRegister.create(Registries.RECIPE_TYPE,
+			CulinaryCompat.MODID);
+
+	public static final RegistryObject<RecipeSerializer<MultiCutting>> MULTI_CUTTING_SERIALIZER = RECIPE_SERIALIZERS
+			.register("multi_cutting", (Supplier<RecipeSerializer<MultiCutting>>) MultiCutting.Serializer::new);
+	public static final RegistryObject<RecipeType<MultiCutting>> MULTI_CUTTING_TYPE = RECIPE_TYPES
+			.register("multi_cutting", () -> new RecipeType<MultiCutting>() {
+				@Override
+				public String toString() {
+					return CulinaryCompat.MODID + ":multi_cutting";
+				}
+			});
+
 	private Recipes() {
 	}